separated out the core functionalities of the backend
